pyproject.toml
[project]
name = "project-9-photo-share"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.117.1,<0.118.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.9,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"
black = "^25.9.0"
ruff = "^0.13.1"

[tool.black]
target-version = ["py311"]
line-length = 88

[tool.ruff]
target-version = "py311"
line-length = 88
indent-width = 4

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E501", "B904"]
